LEGEND: Finish GatewayEngine initPersisted

+++ engines/legend/games/gateway/core/persisted.h
@@ -358,7 +358,7 @@ struct Persisted : public virtual Legend::Early::Persisted {
 	uint16 _val232 = 0;
 	uint32 _val233[4] = { 0 };
 	uint32 _val234[4] = { 0 };
-	uint8 _val235[4] = { 0 };
+	uint16 _val235[4] = { 0 };
 	uint8 _val236 = 0;
 	uint8 _val237 = 0;
 	STRUC20 _val238;
